Broke project classes apart for cleanliness.
