+++ stackgvm/assembler/php/parser/LineKindParser.php
@@ -4,13 +4,20 @@
  * LineKindParser
  *
  */
-class OperandKindParser implements IntegerExpressionParser {
+class LineKindParser implements IntegerExpressionParser {
+
+    const MATCHED = [
+        LineKind::MATCH_INSTRUCTION => LineKind::INSTRUCTION,
+        LineKind::MATCH_LABEL       => LineKind::LABEL,
+        LineKind::MATCH_CODE_SYMBOL => LineKind::CODE_SYMBOL,
+        LineKind::MATCH_DATA_SYMBOL => LineKind::DATA_SYMBOL
+    ];
 
     /**
      * @implements IntegerExpressionParser::parse()
      */
     public function parse(string $sLine) : int {
-        if (empty($sLine) {
+        if (empty($sLine)) {
             return LineKind::BLANK;
         }
         foreach (self::MATCHED as $sMatch => $iKind) {
@@ -18,6 +25,6 @@ class OperandKindParser implements IntegerExpressionParser {
                 return $iKind;
             }
         }
-        throw new ParseException()
+        throw new ParseException("Could not determine line kind '" . $sLine . "'");
     }
 }
